pylzy/requirements.txt
aioboto3>=9.6.0
appdirs>=1.4.4,<2.0.0
azure-storage-blob>=12.9.0,<13.0.0
beartype>=0.10.0,<1.0.0
botocore>=1.27.59
cloudpickle>=2.1.0,<3.0.0
grpcio>=1.46.0
googleapis-common-protos>=1.56.3
importlib-metadata>=4.8.1  # TODO drop after dropping 3.9!
packaging>=21.3.0
pip>=10.0  # we can't use upper boundary here because of the pip release cycle
protobuf>=4.21.1
pycryptodome>=3.15.0,<4.0.0
pyjwt[crypto]>=2.4.0,<3.0.0
pypi-simple>=1.1.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
requests>=2.26.0,<3.0.0
serialzy>=1.2.0,<2.0.0
stdlib-list>=0.8.0,<1.0.0; python_version < '3.10'
get-annotations>=0.1.2; python_version < '3.10'
tqdm>=4.64.0,<5.0.0
typing-extensions>=4.4.0
